requirements.txt
# LLM Providers
anthropic>=0.75.0
openai>=2.14.0

# Document Parsing
pypdfium2>=4.30.0
python-docx>=1.2.0
python-pptx>=1.0.2

# Image Processing
Pillow>=11.0.0

# Web UI
streamlit>=1.40.0
streamlit-antd-components>=0.3.0
streamlit-agraph>=0.0.45

# Configuration
pydantic>=2.12.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0
